Changes to support multiple jobs
Fixing test case and template
